fiancial de 0/bot2.0/backup_dashboard_2025-12-17/requirements.txt
# ============================================
# PROFESSIONAL IOL TRADING BOT v2.0
# Python Dependencies
# ============================================

# Core Python
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Science & Analysis
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
scikit-learn==1.3.2
hmmlearn==0.3.0

# Deep Learning & AI
torch==2.1.2
transformers==4.36.2
stable-baselines3==2.2.1
gymnasium==0.29.1
xgboost==2.0.3
openai==1.6.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# API & Web
requests==2.31.0
aiohttp==3.9.1
fastapi==0.108.0
uvicorn==0.25.0

# Dashboard & Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
loguru==0.7.2
pytrends==4.9.2

# Optimization (NEW)
optuna==3.5.0

# Notifications (NEW)
python-telegram-bot==20.7

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
